doc_to_alt_text/requirements.txt
Django>=4.0
python-docx
requests
